Moving c_misc files into special package
