+++ gamma_inc.go
@@ -5,8 +5,8 @@
 package mathext
 
 import (
-	"github.com/gonum/mathext/internal"
 	"github.com/gonum/mathext/internal/cephes"
+	"github.com/gonum/mathext/internal/special"
 )
 
 // GammaInc computes the incomplete Gamma integral.
@@ -37,7 +37,7 @@ func GammaIncComp(a, x float64) float64 {
 // inclusive or GammaIncInv will panic. GammaIncInv should return a positive
 // number, but can return NaN if there is a failure to converge.
 func GammaIncInv(a, y float64) float64 {
-	return internal.GammaIncInv(a, y)
+	return special.GammaIncInv(a, y)
 }
 
 // GammaIncCompInv computes the inverse of the complemented incomplete Gamma
