+++ internal/special/special.go
@@ -5,7 +5,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package internal
+package special
 
 type objectiveFunc func(float64, []float64) float64
